+++ clikt/src/commonMain/kotlin/com/github/ajalt/clikt/parameters/options/FlagOption.kt
@@ -5,6 +5,7 @@ import com.github.ajalt.clikt.core.BadParameterValue
 import com.github.ajalt.clikt.parameters.types.boolean
 import com.github.ajalt.clikt.parameters.types.int
 import com.github.ajalt.mordant.terminal.YesNoPrompt
+import kotlin.jvm.JvmOverloads
 
 /** A block that converts a flag value from one type to another */
 typealias FlagConverter<InT, OutT> = OptionTransformContext.(InT) -> OutT
@@ -78,9 +79,20 @@ inline fun <OutT> OptionWithValues<Boolean, Boolean, Boolean>.convert(
 
 /**
  * Turn an option into a flag that counts the number of times it occurs on the command line.
+ *
+ * @param limit The maximum number of times the option can be given. (defaults to no limit)
+ * @param clamp If `true`, the counted value will be clamped to the [limit] if it is exceeded. If
+ *   `false`, an error will be shown isntead of clamping.
  */
-fun RawOption.counted(): OptionWithValues<Int, Int, Int> {
-    return int().transformValues(0..0) { it.lastOrNull() ?: 1 }.transformAll { it.sum() }
+@JvmOverloads // TODO(5.0): remove this annotation
+fun RawOption.counted(limit: Int = Int.MAX_VALUE, clamp: Boolean = true): OptionWithValues<Int, Int, Int> {
+    return int().transformValues(0..0) { it.lastOrNull() ?: 1 }.transformAll {
+        val s = it.sum()
+        if (!clamp && s > limit) {
+            fail(context.localization.countedOptionExceededLimit(s, limit))
+        }
+        s.coerceAtMost(limit)
+    }
 }
 
 /**